+++ src/utils.cpp
@@ -10,7 +10,7 @@ std::string  wcstring_to_mbstring(std::wstring const &src)
 
 	size_t new_buf_len = WideCharToMultiByte(CP_UTF8, 0, src.c_str(), -1, NULL, 0, NULL, NULL);
 	char    *multibyte_buffer = new char[new_buf_len];
-	WideCharToMultiByte(CP_UTF8, 0, src.c_str(), -1, multibyte_buffer, new_buf_len, NULL, NULL);
+	WideCharToMultiByte(CP_UTF8, 0, src.c_str(), -1, multibyte_buffer, (int)new_buf_len, NULL, NULL);
 	std::string retval(multibyte_buffer);
 	delete[] multibyte_buffer;
 	return retval;
@@ -24,7 +24,7 @@ std::wstring  mbstring_to_wcstring(std::string const &src)
 	// https://learn.microsoft.com/en-us/windows/win32/api/stringapiset/nf-stringapiset-multibytetowidechar
 	size_t new_buf_len = MultiByteToWideChar(CP_UTF8, 0, src.c_str(), -1, NULL, 0);
 	WCHAR *widechar_buffer = new WCHAR[new_buf_len];
-	MultiByteToWideChar(CP_UTF8, 0, src.c_str(), -1, widechar_buffer, new_buf_len);
+	MultiByteToWideChar(CP_UTF8, 0, src.c_str(), -1, widechar_buffer, (int)new_buf_len);
 	std::wstring retval(widechar_buffer);
 	delete[] widechar_buffer;
 	return retval;
